Updated AI State Control
State Control now uses a pool of player objects to check for players in
vision, rather than simply checking using OnTriggerEnter(). Still need
to get rid of raycasting for vision checks, it doesn't make sense for
games with lots of enemies as the enemies will obstruct the view of
other enemies.