Co-op ai lookin' good
Still needs some cleaning and clear direction on purpose of Player
script.
- added WatchedEnemies to Player script. This is shared between co-op
and ai
- OnTriggerEnter/Exit for watchedEnemies now ensures if other is not a
trigger. The sphere triggers detect each other. CURRENTLY STILL A
PROBLEM ON ENEMY SIDE
- Added ResetOnSwitch() function in AI and Player controllers. When
switching to a player in strategy, this is called to reset all target
and animation info. Then in strategy certain things can be inherited if
chosen.