Revamped Player Controller
Lotsa stuff
Cancel any ability by clicking it again.
Convention in player controller is no longer using bools to track
whether enemyClicked, but rather if targetedEnemy is null.
Zap is now left click on enemy to use. (we need visual indicator of what
is happening) If you are currently shooting an enemy with gun and left
click on something that is not an enemy. it will fire on current enemy.
Player animation is now independent of what action the player is taking.
Now it tries to move towards its special target if it exists, then its
basic target if it exists, then its floor target.