GunFX, Player shoot and walking aim animations
Basic attack is hard-coded to new RifleShot ability
Bugs/weird:
GunFX is still based on projectile hit, I had trouble with setting a
destination vector
Sound is played on impact, not on fire. This is because the particle is
destroyed on impact and will often cut the sound off early.