Start of Revive, need to talk to Josh
Currently player colliders are turned off on death. Just need to check
with Josh if that's permanent since it dealt with enemies. My problem is
we currently can't select a dead player. (I can create a workaround - a
new gameobject collider for dead players).

ActivePlayer can use Revive on non dead teammates to heal them (We may
change the name of Revive later, since I'm thinking it's more of a team
heal/revive).