AI Defend State + small change to tac pause
Press Left Shift and then right click to set ai into defense mode for 8
seconds:
CURRENT BUG: it doesn't work initially on the player you start as. The
coopAI script has to run for at least a frame for the transition to
work.
Claudia: check the commit for the change I did to TacticalPause.cs